Refactor codes and create new files
